Extend grid and background scaling for zoom
